#### What type of PR is this?
/kind feature
/kind api-change

#### What this PR does / why we need it:
Support postfilter plugin

#### Which issue(s) this PR fixes:
Part of #72 

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
```release-note
Support postfilter plugin
```

#### What are the benchmark results of this change?
<!--
If documentation-only or trivial, just write "N/A". Otherwise, run the following:

1. `make bench-example > before.txt` on the commit prior to your changes
2. `make bench-example > after.txt` on the commit of your changes
3. Paste the output of `benchstat before.txt after.txt` as a code block.

If you haven't yet installed benchstat, it looks like this.
```
$ go install golang.org/x/perf/cmd/benchstat@latest
```

-->
```benchstat
goos: darwin
goarch: amd64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e/scheduler
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
                                      │  before.txt  │              after.txt               │
                                      │    sec/op    │    sec/op      vs base               │
Example_NodeNumber/Simple/New-4         198.7m ± 79%    222.6m ± 16%        ~ (p=0.310 n=6)
Example_NodeNumber/Simple/Run-4         475.3µ ± 40%    767.4µ ± 41%  +61.46% (p=0.041 n=6)
Example_NodeNumber/Simple_Log/New-4     189.3m ±  0%    209.4m ± 31%  +10.59% (p=0.015 n=6)
Example_NodeNumber/Simple_Log/Run-4     348.7µ ± 31%    411.6µ ± 18%        ~ (p=0.132 n=6)
Example_NodeNumber/Advanced/New-4       217.4m ±  2%    417.0m ± 34%  +91.86% (p=0.002 n=6)
Example_NodeNumber/Advanced/Run-4       79.20µ ±  1%    95.36µ ± 34%  +20.40% (p=0.002 n=6)
Example_NodeNumber/Advanced_Log/New-4   216.7m ±  1%    257.8m ± 39%  +18.94% (p=0.002 n=6)
Example_NodeNumber/Advanced_Log/Run-4   95.40µ ±  2%   131.44µ ± 18%  +37.77% (p=0.002 n=6)
geomean                                 6.213m          8.169m        +31.50%
```