@Gekko0114 2023-10-28
Hi @sanposhiho ,

Though this is a draft, I have some questions.
Could you answer my comments above?
Thanks,