@Gekko0114 2023-11-19
@sanposhiho ,
Thanks for your comments!
Fixed or replied. PTAL again