@Gekko0114 2023-11-24
Thanks for your review, @sanposhiho !
Fixed it, PTAL again.