@Gekko0114 2023-11-26
@sanposhiho ,
Thanks, fixed it. PTAL